+++ .gitignore
@@ -1,1 +1,2 @@
 nr_agent_sdk-v0.16.1.0-beta.x86_64
+.idea
